Release: v2.50.0

+++ src/detection/localip/localip.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "fastfetch.h"
+#include "modules/localip/option.h"
 
 typedef struct FFLocalIpResult
 {
@@ -20,7 +21,7 @@ typedef struct FFLocalIpNIFlag
     const char *name;
 } FFLocalIpNIFlag;
 
-static inline void ffLocalIpFillNIFlags(FFstrbuf *buf, uint32_t flag, const FFLocalIpNIFlag names[])
+static inline void ffLocalIpFillNIFlags(FFstrbuf *buf, uint64_t flag, const FFLocalIpNIFlag names[])
 {
     for (const FFLocalIpNIFlag *nf = names; flag && nf->name; ++nf)
     {
